datamodules: Switch to kornia AugmentationSequential (#2147)
* Switch to kornia AugmentationSequential in datamodules

* Fix for classification & regression tests

* Fix for segmentation tests

* Revert inria batch_size

* Set keepdim=True

* Fix ssl4eo_benchmark

* Masks must not have a channel dimension

* Fix south africa dimensions

* Fix chesapeake dimensions

* Remove unused import

* mypy fixes & remove _Transform

* Fix quakeset

* mypy fix

* Fix chesapeake plot test

* mypy fix

* Fix SpaceNet aug

* Bump min version of kornia

* Remove conditional squeeze

* Fix for new datasets

* Update pyproject.toml

* Update tutorials

---------

+++ tests/datamodules/test_geo.py
@@ -31,7 +31,7 @@ class CustomGeoDataset(GeoDataset):
         self.res = 1
 
     def __getitem__(self, query: BoundingBox) -> dict[str, Any]:
-        image = torch.arange(3 * 2 * 2).view(3, 2, 2)
+        image = torch.arange(3 * 2 * 2, dtype=torch.float).view(3, 2, 2)
         return {'image': image, 'crs': CRS.from_epsg(4326), 'bounds': query}
 
     def plot(self, *args: Any, **kwargs: Any) -> Figure:
@@ -68,7 +68,7 @@ class CustomNonGeoDataset(NonGeoDataset):
         self.length = length
 
     def __getitem__(self, index: int) -> dict[str, Tensor]:
-        return {'image': torch.arange(3 * 2 * 2).view(3, 2, 2)}
+        return {'image': torch.arange(3 * 2 * 2, dtype=torch.float).view(3, 2, 2)}
 
     def __len__(self) -> int:
         return self.length
